+++ src/__test__/handler.test.ts
@@ -15,7 +15,7 @@ import {
     KnowledgeBaseResult,
     CrmService
 } from "stentor";
-import { CrmResponse } from "stentor-models";
+import { CrmResponse, CrmServiceAvailability } from "stentor-models";
 import { IntentRequestBuilder } from "stentor-request";
 import { ContextBuilder } from "stentor-context";
 
@@ -25,6 +25,16 @@ import { CONTACT_CAPTURE_CURRENT_DATA, CONTACT_CAPTURE_LIST, CONTACT_CAPTURE_SEN
 import { DetailParams, Place, PlacesService, SearchParams } from "../services";
 
 class MockCRM implements CrmService {
+    public update?(): Promise<CrmResponse> {
+        throw new Error("Method not implemented.");
+    }
+    public getAvailability(): Promise<CrmServiceAvailability> {
+        throw new Error("Method not implemented.");
+    }
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    public getJobType(): Promise<any> {
+        throw new Error("Method not implemented.");
+    }
     public async send(): Promise<CrmResponse> {
         return {
             status: "Success"
@@ -1165,6 +1175,7 @@ describe(`${ContactCaptureHandler.name}`, () => {
                         ],
                         transcript: [],
                         refId: undefined,
+                        jobTypeId: undefined,
                         userId: "userId",
                         sessionId: "sessionId",
                         source: "stentor"