+++ src/strategies/FormResponseStrategy.ts
@@ -3,6 +3,7 @@
 import {
     compileResponse,
     Context,
+    CrmService,
     hasSessionId,
     log,
     Request,
@@ -17,6 +18,7 @@ import { isSessionClosed, newLeadGenerationData } from "../utils";
 import { ContactCaptureHandler } from "../handler";
 
 import { ResponseStrategy } from "./ResponseStrategy";
+import { CrmServiceAvailability, SessionStore } from "stentor-models";
 import { MultistepForm } from "../form";
 
 /**
@@ -45,16 +47,16 @@ export interface FormActionResponseData {
 function getContactFormFallback(): Response {
 
     const contactUsForm: MultistepForm = {
-        "name": "contact_us_only",
+        name: "contact_us_only",
         type: "FORM",
-        "header": [
+        header: [
             {
                 "step": "contact_info",
                 "label": "Contact"
             }
         ],
-        "labelHeader": true,
-        "steps": [
+        labelHeader: true,
+        steps: [
             {
                 "crmSubmit": true,
                 "final": true,
@@ -63,43 +65,43 @@ function getContactFormFallback(): Response {
                 "nextAction": "submit",
                 "fields": [
                     {
-                        "name": "full_name",
-                        "label": "Name",
-                        "type": "TEXT",
-                        "mandatory": true
+                        name: "full_name",
+                        label: "Name",
+                        type: "TEXT",
+                        mandatory: true
                     },
                     {
-                        "format": "PHONE",
-                        "name": "phone",
-                        "label": "Phone",
-                        "placeholder": "Your 10 digit phone number",
-                        "type": "TEXT",
-                        "mandatory": true
+                        format: "PHONE",
+                        name: "phone",
+                        label: "Phone",
+                        placeholder: "Your 10 digit phone number",
+                        type: "TEXT",
+                        mandatory: true
                     },
                     {
-                        "name": "message",
-                        "label": "Tell us what you need help with",
-                        "rows": 6,
-                        "type": "TEXT",
-                        "multiline": true
+                        name: "message",
+                        label: "Tell us what you need help with",
+                        rows: 6,
+                        type: "TEXT",
+                        multiline: true
                     }
                 ],
-                "title": "Contact Information"
+                title: "Contact Information"
             },
             {
-                "fields": [
+                fields: [
                     {
-                        "name": "thank_you_text",
-                        "header": {
-                            "title": "Thank You"
+                        name: "thank_you_text",
+                        header: {
+                            title: "Thank You"
                         },
-                        "text": "Somebody will call you as soon as possible.",
-                        "type": "CARD"
+                        text: "Somebody will call you as soon as possible.",
+                        type: "CARD"
                     }
                 ],
-                "previousAction": "omit",
-                "nextAction": "omit",
-                "name": "Thanks"
+                previousAction: "omit",
+                nextAction: "omit",
+                name: "Thanks"
             }
         ]
     };
@@ -170,6 +172,23 @@ function leadSummary(slots: RequestSlotMap, leadDataList: CaptureRuntimeData): s
     return summary;
 }
 
+/**
+ * 
+ * @param busyDays Format this for the widget for "easier" consumption
+ * @returns 
+ */
+function formatBusyDays(busyDays: CrmServiceAvailability): string {
+    const busyDates: string[] = [];
+
+    busyDays.unavailabilities.forEach(value => {
+        if (!value.available) {
+            busyDates.push(value.date.date);
+        }
+    });
+
+    return busyDates.join(",");
+}
+
 export class FormResponseStrategy implements ResponseStrategy {
     public async getResponse(handler: ContactCaptureHandler, request: Request, context: Context): Promise<Response> {
 
@@ -195,6 +214,9 @@ export class FormResponseStrategy implements ResponseStrategy {
 
             // Update the list on session
             context.session.set(Constants.CONTACT_CAPTURE_LIST, leadDataList);
+
+            // First availabilty
+            await this.addAvailability(response, context.services.crmService, context.session);
 
             return response;
         }
@@ -231,12 +253,16 @@ export class FormResponseStrategy implements ResponseStrategy {
 
             // Send the requested form
             if (data.followupForm) {
-                return getFormResponse(handler.data, data.followupForm);
+                response = getFormResponse(handler.data, data.followupForm);
+                await this.addAvailability(response, context.services.crmService, context.session);
+                return response;
             }
 
             // Don't submit until the form says so
             if (!stepFromData.crmSubmit) {
-                return {};
+                response = {};
+                await this.addAvailability(response, context.services.crmService, context.session);
+                return response;
             }
 
             // Form widget has to say we are fininshed (unless session is closed)
@@ -246,6 +272,7 @@ export class FormResponseStrategy implements ResponseStrategy {
         }
 
         const existingRefId = context.session.get(Constants.CONTACT_CAPTURE_EXISTING_REF_ID);
+        const jobType = context.session.get(Constants.CONTACT_CAPTURE_JOB_TYPE);
 
         // if we created a CRM object and then closed, then we are done
         if (isAbandoned && existingRefId) {
@@ -263,6 +290,7 @@ export class FormResponseStrategy implements ResponseStrategy {
             currentUrl: url,
             externalId: hasSessionId(request) ? request.sessionId : "unknown",
             existingRefId,
+            jobTypeId: jobType?.id,
             crmFlags: handler.data?.crmFlags,
             isAbandoned,
         };
@@ -296,6 +324,76 @@ export class FormResponseStrategy implements ResponseStrategy {
 
         context.session.set(Constants.CONTACT_CAPTURE_EXISTING_REF_ID, leadSendResult.id);
 
-        return {}; // form widget - no response
+        // form widget - no response
+        response = {};
+        await this.addAvailability(response, context.services.crmService, context.session);
+        return response;
+    }
+
+    private async addAvailability(response: Response, crmService: CrmService, session: SessionStore): Promise<Response> {
+        const leadDataList: CaptureRuntimeData = session.get(Constants.CONTACT_CAPTURE_LIST);
+
+        let busyDays: CrmServiceAvailability = session.get(Constants.CONTACT_CAPTURE_BUSY_DAYS) as CrmServiceAvailability;
+
+        // First time?
+        if (!busyDays) {
+            busyDays = await crmService.getAvailability({
+                start: null,
+                end: null,
+            });
+
+            session.set(Constants.CONTACT_CAPTURE_BUSY_DAYS, busyDays);
+        } else {
+            // Try to augment if we have a description
+            const messageData = leadDataList.data.find((data) => {
+                return data.slotName?.toLowerCase() === "message";
+            });
+
+            if (messageData?.collectedValue) {
+                const description = messageData.collectedValue?.trim();
+                const existingDescription = session.get(Constants.CONTACT_CAPTURE_DESCRIPTION);
+
+                // Only call if description changed
+                if (description !== existingDescription) {
+                    session.set(Constants.CONTACT_CAPTURE_DESCRIPTION, description);
+
+                    const jobType = await crmService.getJobType(description);
+                    const existingJobType = session.get(Constants.CONTACT_CAPTURE_JOB_TYPE);
+
+                    // Only call if the jobType changed (visitor changed the description)
+                    if (jobType.id !== existingJobType?.id) {
+                        session.set(Constants.CONTACT_CAPTURE_JOB_TYPE, jobType);
+
+                        busyDays = await crmService.getAvailability(
+                            {
+                                start: null,
+                                end: null,
+                            },
+                            {
+                                jobType
+                            }
+                        );
+
+                        session.set(Constants.CONTACT_CAPTURE_BUSY_DAYS, busyDays);
+                    }
+                }
+            }
+        }
+
+        response.context = {
+            active: [
+                {
+                    name: "BusyDays",
+                    timeToLive: {},
+                    parameters: {
+                        busyDays: formatBusyDays(busyDays),
+                    }
+                },
+            ],
+        };
+
+
+        return response;
     }
 }
+